feat(rum-distiller): port to rum-distiller
