+++ packages/spacecat-shared-rum-api-client/package.json
@@ -39,6 +39,7 @@
     "@adobe/helix-shared-wrap": "2.0.2",
     "@adobe/helix-universal": "5.0.5",
     "@adobe/spacecat-shared-utils": "1.4.0",
+    "@adobe/rum-distiller": "1.4.0",
     "aws4": "1.13.2",
     "urijs": "^1.19.11"
   },
