+++ packages/spacecat-shared-rum-api-client/src/functions/opportunities/high-inorganic-high-bounce-rate.js
@@ -10,23 +10,13 @@
  * governing permissions and limitations under the License.
  */
 
-import { DataChunks, generateKey } from '../../common/cruncher.js';
-import { classifyTrafficSource } from '../../common/traffic.js';
+import { DataChunks, facets } from '@adobe/rum-distiller';
+import { loadBundles, trafficSeriesFn } from '../../utils.js';
 
 const HOMEPAGE_PAID_TRAFFIC_THRESHOLD = 0.8;
 const NON_HOMEPAGE_PAID_TRAFFIC_THRESHOLD = 0.5;
 const BOUNCE_RATE_THRESHOLD = 0.5;
 const DAILY_PAGEVIEW_THRESHOLD = 1000;
-
-const trafficSeriesFn = (memo, type) => (bundle) => {
-  const key = generateKey(bundle.url, bundle.id, bundle.time);
-  if (!memo[key]) {
-    // eslint-disable-next-line no-param-reassign
-    memo[key] = classifyTrafficSource(bundle).type;
-  }
-
-  return type === memo[key] ? bundle.weight : 0;
-};
 
 function convertToOpportunity(traffic) {
   const {
@@ -66,9 +56,9 @@ function handler(bundles, opts = {}) {
 
   const dataChunks = new DataChunks();
 
-  dataChunks.loadBundles(bundles);
+  loadBundles(bundles, dataChunks);
 
-  dataChunks.addFacet('urls', (bundle) => bundle.url);
+  dataChunks.addFacet('urls', facets.url);
 
   dataChunks.addSeries('views', (bundle) => bundle.weight);
   dataChunks.addSeries('clicks', (bundle) => (bundle.events.some((e) => e.checkpoint === 'click') ? bundle.weight : 0));
