+++ packages/spacecat-shared-rum-api-client/test/fixtures/cwv.json
@@ -40,11 +40,11 @@
     "pageviews": 1600,
     "lcp": 1111.8000000715256,
     "lcpCount": 9,
-    "cls": null,
+    "cls": 0.00079558170557108,
     "clsCount": 9,
     "inp": 184,
     "inpCount": 6,
-    "ttfb": null,
+    "ttfb": 312.69999980926514,
     "ttfbCount": 9
   },
   {
@@ -172,7 +172,7 @@
     "pageviews": 300,
     "lcp": 1631,
     "lcpCount": 3,
-    "cls": null,
+    "cls": 1,
     "clsCount": 3,
     "inp": 24,
     "inpCount": 2,
@@ -328,9 +328,9 @@
     "pageviews": 100,
     "lcp": 1345.7999999523163,
     "lcpCount": 1,
-    "cls": null,
+    "cls": 0.07786671987900227,
     "clsCount": 1,
-    "inp": null,
+    "inp": 56,
     "inpCount": 1,
     "ttfb": 118.89999997615814,
     "ttfbCount": 1
